requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
dj-database-url==0.5.0
Django==4.2.16
django-allauth==0.57.2
django-crispy-forms==2.3
django-google-maps==0.13.0
django-summernote==0.8.20.0
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
sqlparse==0.5.1
whitenoise==5.3.0
